feat(ch240): group entries on entry form by category
